requirements.txt
mysql-connector-python
pandas
scikit-learn
matplotlib
seaborn
